Move createElement logic to composable
